+++ src/test/kotlin/sparta/nbcamp/wachu/domain/security/JwtTest.kt
@@ -9,8 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
-import sparta.nbcamp.wachu.domain.member.dto.UserRole
-import sparta.nbcamp.wachu.security.jwt.JwtTokenManager
+import sparta.nbcamp.wachu.domain.member.entity.MemberRole
+import sparta.nbcamp.wachu.infra.security.jwt.JwtTokenManager
 import java.nio.charset.StandardCharsets
 
 @SpringBootTest
@@ -29,14 +29,14 @@ class JwtTest {
         val `토큰 값` =
             jwtTokenManager.generateToken(
                 memberId = 1,
-                userRole = UserRole.MEMBER
+                memberRole = MemberRole.MEMBER
             )
 
         val key = Keys.hmacShaKeyFor(secret.toByteArray(StandardCharsets.UTF_8))
 
         val `토큰 꺼내기` = Jwts.parser().verifyWith(key).build().parseSignedClaims(`토큰 값`)
         `토큰 꺼내기`.payload.get("memberId") shouldBe 1
-        `토큰 꺼내기`.payload.get("userRole", String::class.java) shouldBe "MEMBER"
+        `토큰 꺼내기`.payload.get("memberRole", String::class.java) shouldBe "MEMBER"
     }
 
     @Test
@@ -45,11 +45,18 @@ class JwtTest {
         val `토큰 값` =
             jwtTokenManager.generateToken(
                 memberId = 2,
-                userRole = UserRole.ADMIN
+                memberRole = MemberRole.ADMIN
             )
 
         val `검증된 토큰` = jwtTokenManager.validateToken(`토큰 값`)
 
         `검증된 토큰`.isSuccess shouldBe true
+    }
+
+    @Test
+    fun `1`() {
+        val a = 1
+        val b = 2
+        a + b shouldBe 3
     }
 }
